requirements.txt
asyncpg==0.29.0
bcrypt==4.1.3
fastapi==0.111.0
passlib==1.7.4
pydantic==2.7.4
pydantic_settings==2.3.3
python_jose==3.3.0
SQLAlchemy==2.0.31
uvicorn==0.30.1
